api/python-search/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Vector search
pinecone>=4.0.0
sentence-transformers>=2.2.2

# Utilities
python-multipart>=0.0.6 